third_party/blink/web_tests/storage/websql/open-database-creation-callback-expected.txt

CONSOLE ERROR: unable to open database, version mismatch, '1.0' does not match the currentVersion of ''
This test tests openDatabase()'s creation callback.
